requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
blas=1.0=mkl
ca-certificates=2020.1.1=0
certifi=2019.11.28=py37_1
cudatoolkit=10.1.243=h6bb024c_0
freetype=2.9.1=h8a8886c_1
intel-openmp=2020.0=166
joblib=0.14.1=py_0
jpeg=9b=h024ee3a_2
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_0
mkl=2020.0=166
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.0.15=py37ha843d7b_0
mkl_random=1.1.0=py37hd6b4f25_0
ncurses=6.2=he6710b0_0
ninja=1.9.0=py37hfd86e86_0
numpy=1.18.1=py37h4f9e942_0
numpy-base=1.18.1=py37hde5b4d6_1
olefile=0.46=py37_0
openssl=1.1.1e=h7b6447c_0
pillow=7.0.0=py37hb39fc2d_0
pip=20.0.2=py37_1
python=3.7.6=h0371630_2
pytorch=1.4.0=py3.7_cuda10.1.243_cudnn7.6.3_0
readline=7.0=h7b6447c_5
scikit-learn=0.22.1=py37hd81dba3_0
scipy=1.4.1=py37h0b6359f_0
setuptools=46.1.1=py37_0
six=1.14.0=py37_0
sqlite=3.31.1=h7b6447c_0
tk=8.6.8=hbc83047_0
torchvision=0.5.0=py37_cu101
tqdm=4.43.0=py_0
wheel=0.34.2=py37_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
